+++ stars-core/src/main/kotlin/tools/aqua/stars/core/metrics/postEvaluation/FailedMonitorsGroupedByTSCInstanceMetric.kt
@@ -44,6 +44,7 @@ import tools.aqua.stars.core.utils.ApplicationConstantsHolder.CONSOLE_SEPARATOR
  * @param D [TickDifference].
  * @property dependsOn The instance of a [ValidTSCInstancesPerTSCMetric] on which this metric
  *   depends on and needs for its calculation.
+ * @property identifier identifier (name).
  * @property loggerIdentifier identifier (name) for the logger.
  * @property logger [Logger] instance.
  */
@@ -56,7 +57,8 @@ class FailedMonitorsGroupedByTSCInstanceMetric<
     D : TickDifference<D>,
 >(
     override val dependsOn: ValidTSCInstancesPerTSCMetric<E, T, S, U, D>,
-    override val loggerIdentifier: String = "failed-monitors-grouped-by-tsc-instance",
+    override val identifier: String = "failed-monitors-grouped-by-tsc-instance",
+    override val loggerIdentifier: String = identifier,
     override val logger: Logger = Loggable.getLogger(loggerIdentifier),
 ) : PostEvaluationMetricProvider<E, T, S, U, D>, Loggable {